[a11y] Use (invisible) cursor to indicate focus
Although this is just an example application, it should set a good example.
Screen readers heavily rely on the cursor location to let the user know
what is currently selected.  To that end, the putCursor Brick API has been
added to make it possible to put an (invisible) cursor on a (focused) widget.